Revert "Syntactically classify buffers without a workspace"
This reverts commit c25f25282c5609fac7e2525f0d4ab460fd93d44b.
