+++ src/EditorFeatures/CSharpTest/Classification/SyntacticTaggerTests.cs
@@ -1,13 +1,9 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
-using System;
-using System.Collections.Immutable;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.Editor.Implementation.Classification;
 using Microsoft.CodeAnalysis.Editor.UnitTests.Workspaces;
-using Microsoft.CodeAnalysis.Host;
-using Microsoft.CodeAnalysis.Host.Mef;
 using Microsoft.CodeAnalysis.Shared.TestHooks;
 using Microsoft.VisualStudio.Text;
 using Roslyn.Test.Utilities;
@@ -36,19 +32,8 @@ namespace Microsoft.CodeAnalysis.Editor.CSharp.UnitTests.Classification
                     subjectBuffer,
                     workspace.GetService<IForegroundNotificationService>(),
                     AggregateAsynchronousOperationListener.CreateEmptyListener(),
-                    typeMap: null,
-                    taggerProvider: new SyntacticClassificationTaggerProvider(
-                        notificationService: null,
-                        typeMap: null,
-                        viewSupportsClassificationServiceOpt: null,
-                        associatedViewService: null,
-                        allLanguageServices: ImmutableArray<Lazy<ILanguageService, LanguageServiceMetadata>>.Empty,
-                        contentTypes: ImmutableArray<Lazy<ILanguageService, ContentTypeLanguageMetadata>>.Empty,
-                        asyncListeners: ImmutableArray<Lazy<IAsynchronousOperationListener, FeatureMetadata>>.Empty),
-                    viewSupportsClassificationServiceOpt: null,
-                    associatedViewService: null,
-                    editorClassificationService: null,
-                    languageName: null);
+                    null,
+                    new SyntacticClassificationTaggerProvider(null, null, null));
 
                 SnapshotSpan span = default(SnapshotSpan);
                 tagComputer.TagsChanged += (s, e) =>
